requirements.txt
crewai
gradio

python-dotenv
requests
